Auth To Journal End points

+++ src/main/java/com/example/alpha/Service/JournalEntryService.java
@@ -30,7 +30,7 @@ public class JournalEntryService {
             JournalEntry saved =  journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
             // if we some error occur then journal entry save but kisi user se asscociate nahi rahegi
-            userService.saveEntry(user);
+            userService.saveUser(user);
         } catch (Exception e){
             throw new RuntimeException("An error occured while saving the entry",e);
         }
@@ -49,11 +49,23 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x->x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x->x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+
+        } catch (Exception e){
+            throw new RuntimeException("An error occurence while saving the entry",e);
+        }
+        return removed;
+
+
     }
 
     public JournalEntry updateEntry(ObjectId id, JournalEntry updatedEntry) {
@@ -68,4 +80,8 @@ public class JournalEntryService {
 
         return journalEntryRepository.save(existing);
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
 }
